Move from apsystems_local to apsystems domain
