+++ .strict-typing
@@ -84,7 +84,7 @@ homeassistant.components.api.*
 homeassistant.components.apple_tv.*
 homeassistant.components.apprise.*
 homeassistant.components.aprs.*
-homeassistant.components.apsystems_local.*
+homeassistant.components.apsystems.*
 homeassistant.components.aqualogic.*
 homeassistant.components.aquostv.*
 homeassistant.components.aranet.*
