+++ homeassistant/components/apsystems/coordinator.py
