+++ homeassistant/components/apsystems/manifest.json
@@ -1,10 +1,10 @@
 {
-  "domain": "apsystems_local",
+  "domain": "apsystems",
   "name": "APsystems",
   "codeowners": ["@mawoka-myblock", "@SonnenladenGmbH"],
   "config_flow": true,
   "dependencies": [],
-  "documentation": "https://www.home-assistant.io/integrations/apsystems_local",
+  "documentation": "https://www.home-assistant.io/integrations/apsystems",
   "homekit": {},
   "iot_class": "local_polling",
   "requirements": ["apsystems-ez1==1.3.1"],
