+++ homeassistant/components/apsystems/strings.json
