+++ mypy.ini
@@ -601,7 +601,7 @@ disallow_untyped_defs = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.apsystems_local.*]
+[mypy-homeassistant.components.apsystems.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
