+++ requirements_all.txt
@@ -457,7 +457,7 @@ apprise==1.7.4
 # homeassistant.components.aprs
 aprslib==0.7.2
 
-# homeassistant.components.apsystems_local
+# homeassistant.components.apsystems
 apsystems-ez1==1.3.1
 
 # homeassistant.components.aqualogic
