+++ tests/components/apsystems/conftest.py
@@ -10,7 +10,7 @@ import pytest
 def mock_setup_entry() -> Generator[AsyncMock, None, None]:
     """Override async_setup_entry."""
     with patch(
-        "homeassistant.components.apsystems_local.async_setup_entry",
+        "homeassistant.components.apsystems.async_setup_entry",
         return_value=True,
     ) as mock_setup_entry:
         yield mock_setup_entry
